+++ packages/next/src/data/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from './usePost';
 export * from './usePosts';
 export * from './useSearch';
+export * from './useSearchNative';
 export * from './useAppSettings';
 export * from './useMenu';
 export * from './useTerms';