+++ wp/headless-wp/includes/classes/BaseToken.php
@@ -52,16 +52,24 @@ abstract class BaseToken {
 		// No secure auth key found. Throw an error.
 		$error = new \WP_Error(
 			'no-secure-auth-key',
-			'Please define either SECURE_AUTH_KEY or TENUP_HEADLESS_JWT_AUTH_KEY in your wp-config.php file.'
+			__( 'Please define either SECURE_AUTH_KEY or TENUP_HEADLESS_JWT_AUTH_KEY in your wp-config.php file.', 'headless-wp' )
 		);
 
-		throw new \Exception( $error->get_error_message() );
+		throw new \Exception(
+			wp_kses_post(
+				$error->get_error_message()
+			)
+		);
 	}
 
 	/**
 	 * Decode capability tokens if present.
+	 *
+	 * @param string $token if set will use this instead of reading from the headers
+	 *
+	 * @return null|object
 	 */
-	protected static function get_payload_from_token() {
+	public static function get_payload_from_token( $token = '' ) {
 		// Get HTTP Authorization Header.
 		$header = isset( $_SERVER['HTTP_AUTHORIZATION'] )
 		? sanitize_text_field( wp_unslash( $_SERVER['HTTP_AUTHORIZATION'] ) )
@@ -72,22 +80,29 @@ abstract class BaseToken {
 			$header = sanitize_text_field( wp_unslash( $_SERVER['REDIRECT_HTTP_AUTHORIZATION'] ) );
 		}
 
+		if ( ! $header && isset( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) ) {
+			$header = sanitize_text_field( wp_unslash( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) );
+		}
+
 		// No Authorization Header is present.
-		if ( ! $header ) {
+		if ( ! $header && ! $token ) {
 			return null;
 		}
 
 		// Get and parse the token.
 		try {
-			list( $token ) = sscanf( $header, 'Bearer %s' );
+			if ( ! $token ) {
+				list( $token ) = sscanf( $header, 'Bearer %s' );
+			}
 
 			if ( empty( $token ) ) {
 				return null;
 			}
+
 			$payload = JWT::decode(
 				$token,
 				self::get_private_key(),
-				array( 'HS256' )
+				[ 'HS256' ]
 			);
 		} catch ( \Exception $e ) {
 			// Token is not valid.