+++ wp/headless-wp/plugin.php
@@ -3,7 +3,7 @@
  * Plugin Name: HeadstartWP
  * Plugin URI:  https://github.com/10up/headstartwp-plugin
  * Description: Adds functionality to the WordPress admin and REST API for 10up's headless framework.
- * Version: 1.0.13
+ * Version: 1.1.3
  * Author:      10up
  * Author URI:  https://10up.com
  * Text Domain: headstartwp
@@ -17,7 +17,7 @@ namespace HeadlessWP;
 use HeadlessWP\Preview\PreviewToken;
 
 // Useful global constants.
-define( 'HEADLESS_WP_PLUGIN_VERSION', '1.0.13' );
+define( 'HEADLESS_WP_PLUGIN_VERSION', '1.1.3' );
 define( 'HEADLESS_WP_PLUGIN_URL', plugin_dir_url( __FILE__ ) );
 define( 'HEADLESS_WP_PLUGIN_PATH', plugin_dir_path( __FILE__ ) );
 define( 'HEADLESS_WP_PLUGIN_INC', HEADLESS_WP_PLUGIN_PATH . 'includes/' );
@@ -40,17 +40,17 @@ if ( file_exists( HEADLESS_WP_PLUGIN_PATH . '/vendor/autoload.php' ) ) {
 	require_once HEADLESS_WP_PLUGIN_PATH . 'vendor/autoload.php';
 } else {
 	spl_autoload_register(
-		function( $class ) {
+		function ( $the_class ) {
 				// Project-specific namespace prefix.
 				$prefix = 'HeadlessWP\\';
 				// Base directory for the namespace prefix.
 				$base_dir = __DIR__ . '/includes/classes/';
 				// Does the class use the namespace prefix?
 				$len = strlen( $prefix );
-			if ( strncmp( $prefix, $class, $len ) !== 0 ) {
+			if ( strncmp( $prefix, $the_class, $len ) !== 0 ) {
 				return;
 			}
-				$relative_class = substr( $class, $len );
+				$relative_class = substr( $the_class, $len );
 				$file           = $base_dir . str_replace( '\\', '/', $relative_class ) . '.php';
 				// If the file exists, require it.
 			if ( file_exists( $file ) ) {
@@ -66,7 +66,7 @@ $plugin->register();
 
 add_action(
 	'rest_api_init',
-	function() {
+	function () {
 		PreviewToken::setup();
 	}
 );